Added Product Loction and company Location feature. Added Product Identification code

+++ database/migrations/2017_04_17_080333_create_map_locations_table.php
@@ -17,8 +17,8 @@ class CreateMapLocationsTable extends Migration
             $table->increments('id');
             $table->integer('company_id')->unsigned();
             $table->foreign('company_id')->references('id')->on('companies');
-            $table->integer('lon');
-            $table->integer('lat');
+            $table->double('lon');
+            $table->double('lat');
             $table->timestamps();
         });
     }
